+++ tests/test_qscore.py
@@ -1,12 +1,15 @@
 """Tests for Qscore estimation"""
 
 from iqm.benchmarks.optimization.qscore import *
-
-
-backend = "IQMFakeAdonis"
+from iqm.qiskit_iqm.fake_backends.fake_adonis import IQMFakeAdonis
+from iqm.qiskit_iqm.fake_backends.fake_deneb import IQMFakeDeneb
 
 
 class TestQScore:
+    backend = IQMFakeAdonis()
+    qpu_topology = "crystal"
+    custom_qubits_array = [[0], [0, 2], [0, 1, 2], [0, 1, 2, 3], [0, 1, 2, 3, 4]]
+
     def test_qscore(self):
         EXAMPLE_QSCORE = QScoreConfiguration(
             num_instances=2,
@@ -18,11 +21,17 @@ class TestQScore:
             use_virtual_node=True,
             use_classically_optimized_angles=True,
             choose_qubits_routine="custom",
-            custom_qubits_array=[[2], [2, 0], [2, 0, 1], [2, 0, 1, 3], [2, 0, 1, 3, 4]],
+            custom_qubits_array=self.custom_qubits_array,
             seed=1,
             REM=True,
             mit_shots=10,
+            qpu_topology=self.qpu_topology
         )
-        benchmark = QScoreBenchmark(backend, EXAMPLE_QSCORE)
+        benchmark = QScoreBenchmark(self.backend, EXAMPLE_QSCORE)
         benchmark.run()
         benchmark.analyze()
+
+class TestQScoreDeneb(TestQScore):
+    backend = IQMFakeDeneb()
+    qpu_topology = "star"
+    custom_qubits_array = [[1], [1, 2], [1, 2, 3], [1, 2, 3, 4], [1, 2, 3, 4, 5], [1, 2, 3, 4, 5, 6]]